cache opf/cha paths
It does not take too long but it is not quick either. There is very little probability that the paths would change while the program is running. It is unlikely to change between program runs as well so I could just save two lists of paths to a folder. Right now, the path lists are save to a separate repository. Referencing that separate repository seems like a strange solution.