fix: cha exporting function should throw an error if it could not export some chas
This is still some silly silly code: the idea was to use the return value as an indicator that something went wrong, so it was supposed to return the path to the log only if there were some errors/problems. But it also meant that if there were some errors/problems, the function returned so the code with warnings for problems and exception for the errors was unreachable.

This is still a bad idea and this should be handled totally differently. Later. Some day.