feat: allow skipping checking for FIXMEs when scattering all files at once
This useful, for example, when restoring from a backup.
There is a good chance that you are restoring because you messed up the current versions - no need to back up them.

Also added a message telling the user what to do when there are FIXMEs to fix with the code for them to run.

The logic is also a bit clearer now:
- if you want to check for FIXMEs
 - create the missing_*.csv files,
 - do not copy anything if there are FIXMEs to fix
- if you don't want to check:
 - then don't check at all,
 - just copy