fix: remove subregions without annotations only before they have been split
During the processing of regions, a subregion might be split in two because, for example, it has a skip region in the middle. If any one of the two resulting parts, or even the skip region, has annotations in them, we should count the whole subregion as listened to, not just that one part.
Therefore, we should only count annotations before such splitting.