fix(merge basic level): false positive on duplicate annotation ids
Merging of a new sparse_code csv with the old one is done based on annotation ids: each new one is looked up in the old file. Some rows, however, do not have annotation ids (only comments, I think) and so each of them matches a number of row in the old csv which also have an empty annotation id. This was falsely flagged as duplication of annotation ids in the old file.