fix: put __init__.py in data folders
I forgot that the data discovery method that is used to find etf_templates in blabpy.vihi.intervals.templates relies on the folder with the etf templates to be a package.
Thus, "__init__.py" is necessary (see the previous commit message for the explanation, see its contents to learn that I deleted that file in that previous commit).