remove the cli script test
It would be too much work to test it while ensuring that no live files are touched.
A way to do that is to:
- create a fake path to a recording folder from tmp_path as if it was the pn-opus root
- use subprocess.Popen or friends and set PN_OPUS_PATH to tmp_path/PN_OPUS

That should probably work but could still easily go wrong if future code does not use or prioritize PN_OPUS_PATH env var when lookin for pn-opus.