refactor: convert some parameters to constants
It reduces the risk of changing parameters that are meant to be held constant in the middle of a sequence of processing stages.
It would never make sense to use one set of context offsets when creating eaf, and another - when creating a report of added intervals.
Similarly, there is no reason for sampling frequency and filtering boundaries to change within the `compute_energy_loudness`.

Also:
Satisfied PyCharm by changing some of the variables that shadowed names from outer scope.