fix: respect the current format of selected regions df
That is the log that is written alongside the eaf file.
It has a limited number of columns: "id", "clip_num", "onset", "offset" with the last two referring to the context tier.
Let's not write anything different quite yet.

This table was reconstructed from the full recording id and the list of context intervals.
Well, the only thing left to reconstruct is "clip_num", which is the value of the "code_num" tier.
Since it was always 15 new intervals, it would get populated by numbers from 1 to 15.
Now that we add to the existing intervals, this won't work.
So, instead a list of code_num now must be supplied to `create_selected_regions_df`.
For the random interval sampling, it is still just the sequence 1 to 15, the only thing is that it is now constructed before calling `create_selected_regions_df`.

Plus a type fixed in "pipeline.py:90": calucalted -> calculated.