allow skipping sparse_code_csv backups
Sometimes, "scattering" sparse_code_csvs will break in the middle and half the files will have a backup created and half will not.
In that case, it is ok to not try to backup the already backed up files again.
If there are multiple changes made in one day, this will result in only one backup made.
Since we are saving sparse_code_csvs before any updates anyway, one backup per day is more than enough.