extract anonymized (sub)recordings info
If there are pauses in LENA recordings, LENA still outputs one big wav file.
Since we use wav-based timing for tokens in seedlings_nouns.csv, potential pauses in the recording are masked, it all looks like one long recording.
We need the timing info of (sub)recordings to be able to reconstruct the timeline in terms of wall clock time.