separate progress bars for audio and video
Video recordings are processed instantaneously, audio recordings take 1-5 seconds to process depending on the connection speed.
Having them all under one progress bar resulted in very unrealistic ETAs.
If video was processed first then the ETA would be seconds and then it would linearly converge to the actual total time - not very useful.
If audio was processed first then the ETA would be twice as what it should be because once the audio was done (50%), the whole process was done.
ETA was half as what it should have been.

Two progress bar are also a bit strange - one takes time to finish, the other one doesn't.
I wasn't able to figure out a better solution.