remove path arguments from make_updated_seedlings_nouns
It's not a great idea to make the user define all the variables.
Now, make_updated_seedlings_nouns will make assumptions about the paths and fail if they are wrong.
The version with the arguments is still available as _make_updated_seedlings_nouns.

Extra: added function to locate BLAB_DATA and seedling-nouns_private within it.