fill in ends of last surpluses in 06-07
For these months, the last surplus region starts at the last top-4 region and finishes at the end of the recording.
I implemented that part before implementing finding recordings' duration, so the end was left as NA signifying "however long till the end of the recording".
This should all be refactored. This whole thing should be refactored.