rename param to exist_identical_ok
Functions that add elements to an EAF will raise an error if an element of a given type with a given id already exists - by default.
This error can be ignored if the element has the same type (tag), id, all other attributes, and text.
The corresponding function parameter is now called exist_identical_ok.

Note: Testing that elements are "identical" is shallow: the subelements are not checked.