fix: correct remote name is 'vihi_main'
At an earlier stage, the plan was to clone/push from/to GitHub and the remote name incorrectly reflected that.

Working with GitHub directly would require us to give all RAs access to the GitHub repo which could potentially be dangerous as RAs are usually less knowledgeable in all things git.