refactor: use "else:" clause for clarity
I had `break` in the "if" clause, so the rest of the code in the loop would only run if the "if" part didn't.
In this case, it was confusing because the code in both clauses is doing similar things, it isn't just "if True then don't run the code below".
Anyway, I think it is clearer this way.

Extra: finish an unfinished comment.