add: pyprojroot to requirements
Useful for locating files within a project. It is similar to pyhere but more flexible in that it can look for the project root using a specific criterion. So, if I have a requirements.txt in a subfolder of a git repo, I can still find the repo and not the subfolder.