fix: convert ids to list before concatenating with another list
fund_child_annotation_ids incorrectly returned back parent annotation ids, only as a list, not as a series.
So when I concatenated parent and child annotations, what I actually did was concatenate a series of strings and a same-length and same-values list of string.
The output of that operation is a series with all string doubled ("a52" -> "a52a52").