fix: close git.Repo objects
If the repo isn't closed, Windows will keep references to temporary files and won't allow to delete the folder_to_clone_into which we need when we clone to a temporary folder first. See help for git.Repo.close() for details. I am not sure if the files are temporary because git creates them that way, gitpython creates them that way, or because the temporary folder was created using tempfile.TemporaryDirectory().

Likely unrelated to Lilli's error because the solution is supposed to be needed only on Windows. But who knows, maybe the setup of pn-opus leads to the same error.