handle the special case of Audio_45_10 sub-recordings
The first recording was a failure. The second one was done 11 days later with a recorder that already had 10 hours on it. Since the recorder can only record for 16 hours, only ~5.5 hours of usable data was recorded. The .wav and .cha files do reflect that. The .its, however, corresponds to the full 16 hours as it includes the first ~10-hour-long recording information. With this, we are dropping that first sub-recording.

Note: the .its file was previously trimmed too but that version was lost.