feat: slightly more precise audio durations
For audio, durations were calculated as the sum of sub-recording durations which were, in turn, calculated as the difference between the clock date-times extracted from the .its and those didn't contain ms in them. I've added a column that has the end of each sub-recording as the position in the wav file - in ms. I've also renamed the columns: what used to be start, end, start_postion_ms is now start_dt, end_dt, start_ms and the new column is called end_ms.

The function that extracts the sub-recording info is now called Its.gather_sub_recordings and also returns the four columns described above. Its code now also use the word *sub*-recordings for consistency with the seedlings-nouns code. Previous naming (just recordings) was chosen to be consistent with rlena which isn't that useful here.

I also had to update the amended sub-recordings for 45_10. I am also committing the script that I used to do that, in case we need to amend more files.