fix: determine last surplus end right away for 06-07
It used to be updated in `_post_process_regions` which was called after we had already calculated the surplus time for each audio in `gather_recording_nouns_audio`. That meant that the last surplus region was effectively dropped from the calculation.

Note: I hate pandas for just ignoring NAs when you calculate sum(). If it didn't, `surplus_ms` would be NA for 06-07 and I would have noticed this bug right away.

Surplus regions are created as complements to top 4 and silence. When you create the complement regions, the first and the last regions will have a missing boundary which need to be set to 0 and the recording duration. Previously, I updated the last region's boundary when the whole corpus was processed already but now I moved duration calculation to the top of `gather_recording_nouns_audio` and then passed it to `get_top3_top4_surplus_regions` so that it could be used right away.