fix: trim regions that end after the recording does
This happened because when we delineated subregions, we incorrectly treated intervals in 5min.csv's as corresponding to 5 minutes of the recordings. In reality, they corresponded to 5 minutes of clock time starting at multiples of 5 minutes after the hour (00, 05, 10, 15, ...). So, the intervals that contained the boundaries of sub-recordings were shorter, potentially up to almost full 5 minutes. For most of the recordings that contained just one uninterrupted 16-hour-long sub-recording, this resulted in exactly 5 extra minutes.

Because subregions were 1 hour long, we decided to ignore the whole thing and not realign subregions after the fact. Especially since that might have entailed not just moving them a bit but also selecting different set of 12 5-minute intervals altogether.