fix(par.ann.): deal with connection issues
- If pushing fails, try again with adjusted parameters.
- If it fails again, print a message but let `vihi annotation start` (more specifically, not error out when it calls `sparse_clone`) finish.

The error that we got was 'fatal: the remote end hung up unexpectedly'. The internet and Claude say that it must be a connection issue. Trying some of the suggested parameters didn't work so we are going to try to apply them all at the same time and, more importantly, we won't error out on a failed push - we can always push later.