fix: avoid reusing annotation IDs
Previously, new IDs were calculated from as the maximum currently used value plus one. In case the annotation with the current maximum value is removed, the new annotation will get its ID which we want to avoid. Now, we will take the greater of the maximum of the current annotation IDs and the last used annotation ID as preserved in the corresponding XML property.