fix: update time slot values, not refs
When an annotation was moved in THEIRS (and possibly had its value changed) but not in OURS, we should update the time values of the time slots as referenced in OURS, not copy the references from THEIRS.
Expecting copying of the time reference slots was dumb anyway (my bad!) but it turned out that additionally ELAN will regularly reassing time slot IDs so even if I tried merging time slots (say, in _disambiguate_added_ids), it would have still led to a wrong result.
Oh well. I am just glad I caught it now.